Make sure type definitions are complete
We have to use the ES6 module system to indicate to TypeScript that
variables are scoped to each file, and hence that e.g. rdflib can
be imported twice in different files.

Furthermore, only ES6 exports are listed in the type definitions,
so I've added a default export in addition to the existing
module.exports. This is a bit idiosyncratic, since the actual
import will probably use the module.exports import (using
TypeScript's `allowSyntheticDefaultImports` option), but it doesn't
really do harm either.