Restore undefined newThing
Note that it is also present in `buttons`, but since it was just
two lines and does not appear to be related specifically to
buttons, I went with just duplicating the code for forms.

It was broken initially by
f618510e95d139fceeb0bbda9e85c50239160090, in which `newThing` was
moved to buttons but not to forms.