Do not add an "Add" button when modifying
I really have no idea when `doingDefaults` is supposed to be true,
 but given that these lines were introduced in a commit that said
it was "basically" working
(2e63674702f6c08cef57e771c06f87dbe0a1a0c8), I hope Tim can clarify
that during review. I _think_ it means when we're not rendering
folder-specific sharing settings, in which case we do not want to
add yet another "Add" button - it was already added when loading
defaults. If this commit is in `master`, it has been approved ;)