Migrating ACL files to TypeScript
Showcasing how migration can be done.

Note that I had to do cast a lot of types because of previously untyped variables that can be various types. We should avoid these ambiguous variables in the future, as it becomes difficult to read and can lead to some surprising results.

I also had to throw a HTMLElement to any (line 91 in aclControl.ts) as code later on heavily rely on the ability to add properties to this variable. This is also something we should avoid, as using undocumented properties might open up for bugs later on.