Refactoring to improve readability
This is huge refactor (restructuring code without changing external behavior).

With this I hope to improve readability of the code, to make it easier for newcomers to come in to the code and be able to change code without fear of not understanding the code properly.

Note that there is still improvements to be done wrt some of the uses of scopes (they are very deep at times, and state is difficult to follow), but I didn't want to risk breaking any code by changing this behavior to much.