Changing how mocking of rdflib works + fixing most tests
This allows for reusing most of rdflib.js as is, with the exception of Fetcher and UpdateManager (the latter is dependent on the former).

I've fixed most tests that broke, with exception of peoplePicker.test.ts, pad.test.ts and forms.test.ts, as these are being currently worked on by others. I've simply commented this out, and they should prioritize their own work when merge conflict happen, and then fix the broken tests that may have come.

Note: We might want to change UpdateManager in rdflib to optionally take in Fetcher so that we don't need to mock it.