Support login from Data Kitchen
The box.refresh method is one place where we can NOT use authn.currentUser() because what we want to know is not just is the user authorized to be here but are they actually logged in.  So we need authn.checkUser() here.  Otherwise when we are in Data-Kitchen we appear to be logged in to the local host  (so we can access it all) but we aren't actually logged in to it.  However we want the option to login to other places so box.refresh needs to check whether we are actually logged in somewhere else and if not give us the issuer popup.